spans for custom apps
