+++ docs/trulens_eval/api/trace/index.md
@@ -1,0 +1,3 @@
+# Trace
+
+::: trulens_eval.trace
