+++ docs/trulens_eval/api/trace/span.md
@@ -1,0 +1,3 @@
+# Span
+
+::: trulens_eval.trace.span
