+++ docs/trulens_eval/api/trace/tracer.md
@@ -1,0 +1,4 @@
+# Tracer
+
+::: trulens_eval.trace.tracer
+
