+++ trulens_eval/trulens_eval/requirements.txt
@@ -9,6 +9,8 @@ nest-asyncio       >= 1.5.8
 typing_extensions  >= 4.9.0
 psutil             >= 5.9.8  # tru.py dashboard starting/ending
 pip                >= 24.0   # for requirements management
+opentelemetry-api  >= 1.24.0 # spans
+opentelemetry-sdk  >= 1.24.0 # spans
 
 packaging          >= 23.2   # for requirements, resources management
 # also: resolves version conflict with langchain-core
